chore(test): explicit .toByte() to silence Copilot R6 false-positive (BAT-1021)
Copilot R6 claimed `ByteArray(64) { 0x7F }` won't compile because the
lambda returns Int not Byte. This is a false positive — Kotlin's
constant-integer-literal conversion accepts integer constants in
contexts expecting Byte/Short when they fit in range. 0x7F = 127 =
Byte.MAX_VALUE; the original line compiled and ran fine across
20/20 tests in 5 prior rounds and multiple BUILD SUCCESSFUL passes.

Adding `.toByte()` explicitly to silence the comment and avoid
re-iteration cycles. Pure cosmetic. 20/20 still pass.

+++ app/src/test/java/com/seekerclaw/app/solana/SolanaWalletManagerTest.kt
@@ -163,7 +163,7 @@ class SolanaWalletManagerTest {
         )
         fake.queueFailure(message = "Auth token invalid", e = staleExc)
         // Retry: Success with fresh token T2 and a signature payload
-        val sigPayload = ByteArray(64) { 0x7F }
+        val sigPayload = ByteArray(64) { 0x7F.toByte() }
         fake.queueSuccess(token = "T2", payload = sigPayload)
 
         val result = SolanaWalletManager.signAndSendTransactionInternal(
